pyproject.toml
[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
    ".git",
    "__pycache__",
    ".pytest_cache",
    ".venv",
    "venv",
    "build",
    "dist",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-live-demo"
version = "0.1.0"
description = "Gemini Live Demo"
requires-python = "==3.13.9"
dependencies = [
    "livekit-agents>=1.3.6",
    "livekit-plugins-google>=1.3.6",
    "python-dotenv>=1.0.1",
    "google-cloud-secret-manager>=2.21.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "ty>=0.0.1a9",
]

[tool.ty]
# Basic ty configuration
